+++ ragavi/ragavi.py
@@ -22,16 +22,15 @@ from dask import delayed, compute
 from datetime import datetime
 from future.utils import listitems, listvalues
 
+from bokeh.events import PlotEvent
 from bokeh.io import (export_png, export_svgs, output_file, output_notebook,
                       save, show)
 from bokeh.layouts import row, column, gridplot, widgetbox, grid
 from bokeh.models import (BasicTicker, CheckboxGroup, ColumnDataSource,
                           CustomJS, HoverTool, Range1d, Legend, LinearAxis,
                           PrintfTickFormatter, Select, Slider, Text, Title,
-                          Toggle)
-from bokeh.models.markers import (Circle, CircleCross, Diamond, Hex,
-                                  InvertedTriangle, Square, SquareCross,
-                                  Triangle)
+                          Toggle, Whisker)
+
 from bokeh.models.widgets import Div, PreText
 from bokeh.plotting import figure
 
@@ -55,8 +54,8 @@ NB_RENDER = None
 BATCH_SIZE = 16
 
 
-logger = vu.config_logger()
-sys.excepthook = vu._handle_uncaught_exceptions
+logger = vu.logger
+excepthook = vu._handle_uncaught_exceptions
 
 #######################################################################
 #################### Define some data Processing class ################
@@ -315,7 +314,7 @@ class DataCoreProcessor:
         if flag:
             # if flagging is activated select data only where flag mask is 0
             processed = self.compute_ydata(ydata, yaxis=yaxis)
-            y = processed.where(flags < 1)
+            y = processed.where(flags == False)
         else:
             y = self.compute_ydata(ydata, yaxis=yaxis)
 
@@ -361,8 +360,10 @@ class DataCoreProcessor:
 
             # shorting y2 to y1 to avoid problems during plotted
             # y2 does not exist for this table
-            d = Data(x=prepd_x, x_label=xlabel, y1=y1, y1_label=y1_label,
-                     y1_err=y1_err, y2=y1, y2_label=y1_label, y2_err=y1_err)
+            d = Data(x=prepd_x, x_label=xlabel, y1=y1,
+                     y1_label=y1_label, y1_err=y1_err,
+                     y2=y1, y2_label=y1_label,
+                     y2_err=y1_err)
 
             return d
 
@@ -401,8 +402,10 @@ class DataCoreProcessor:
                                                   y1_err.data, y2.data,
                                                   y2_err.data)
 
-        d = Data(x=prepd_x, x_label=xlabel, y1=y1, y1_label=y1_label,
-                 y1_err=y1_err, y2=y2, y2_label=y2_label, y2_err=y2_err)
+        d = Data(x=prepd_x, x_label=xlabel, y1=y1,
+                 y1_label=y1_label, y1_err=y1_err,
+                 y2=y2,
+                 y2_label=y2_label, y2_err=y2_err)
 
         return d
 
@@ -527,8 +530,7 @@ def determine_table(table_name):
         return -1
 
 
-def errorbar(fig, x, y, xerr=None, yerr=None, color='red', point_kwargs={},
-             error_kwargs={}):
+def errorbar(fig, x, y, yerr=None, color='red'):
     """Function to plot the error bars for both x and y.
        Takes in 3 compulsory parameters fig, x and y
 
@@ -539,51 +541,34 @@ def errorbar(fig, x, y, xerr=None, yerr=None, color='red', point_kwargs={},
         x_axis value
     y: numpy.ndarray
         y_axis value
-    xerr: numpy.ndarray
-        Errors for x axis, must be an array
     yerr: numpy.ndarray
         Errors for y axis, must be an array
     color: str
         Color for the error bars
 
-
     Outputs
     -------
-    h: fig.multi_line
-        Returns a multiline object for external legend rendering
+    ebars: Whisker annotation model object
+          Return the object containing errorbars
 
     """
     # Setting default return value
-    h = None
-
-    if xerr is not None:
-
-        x_err_x = []
-        x_err_y = []
-
-        for px, py, err in zip(x, y, xerr):
-            x_err_x.append((px - err, px + err))
-            x_err_y.append((py, py))
-
-        h = fig.multi_line(x_err_x, x_err_y, color=color, line_width=3,
-                           level='underlay', visible=False, **error_kwargs)
+    ebars = None
 
     if yerr is not None:
-        y_err_x = []
-        y_err_y = []
 
-        for px, py, err in zip(x, y, yerr):
-            y_err_x.append((px, px))
-            y_err_y.append((py - err, py + err))
+        src = ColumnDataSource(data=dict(upper=y + yerr,
+                                         lower=y - yerr, base=x))
 
-        h = fig.multi_line(y_err_x, y_err_y, color=color, line_width=3,
-                           level='underlay', visible=False, **error_kwargs)
+        ebars = Whisker(source=src, base='base', upper='upper',
+                        lower='lower', line_color=color, visible=False)
+        ebars.upper_head.line_color = color
+        ebars.lower_head.line_color = color
+        fig.add_layout(ebars)
+    return ebars
 
-    return h
 
-
-def make_plots(source, ax1, ax2, fid=0, color='red', y1_err=None,
-               y2_err=None):
+def make_plots(source, ax1, ax2, fid=0, color='red', y1err=None, y2err=None):
     """Generate a plot
 
     Inputs
@@ -599,9 +584,9 @@ def make_plots(source, ax1, ax2, fid=0, color='red', y1_err=None,
          field id number to set the line width
     color: str
         Data points' color
-    y1_err: numpy.ndarray
+    y1err: numpy.ndarray
         y1 error data
-    y2_err: numpy.ndarray
+    y2err: numpy.ndarray
         y2 error data
 
     Outputs
@@ -610,32 +595,57 @@ def make_plots(source, ax1, ax2, fid=0, color='red', y1_err=None,
         Tuple of glyphs
 
     """
-    markers = [Circle, Diamond, Square, Triangle, InvertedTriangle, Hex]
+    markers = ['circle', 'diamond', 'square', 'triangle',
+               'hex']
+    fmarker = 'inverted_triangle'
     glyph_opts = {'size': 4,
                   'fill_alpha': 1,
                   'fill_color': color,
-                  'line_width': 0,
-                  'line_color': 'black'}
+                  'line_color': 'black',
+                  'nonselection_fill_color': '#7D7D7D',
+                  'nonselection_fill_alpha': 0.3}
 
-    nonsel_glyph = markers[fid]()
-    nonsel_glyph.update(fill_color='#7D7D7D',
-                        fill_alpha=0.3,)
+    # if there is any flagged data enforce fmarker where flag is active
+    fmarkers = None
+    if np.any(np.isnan(source.data['y1'])):
+        fmarkers = gen_flag_data_markers(source.data['y1'], fid=fid,
+                                         markers=markers, fmarker=fmarker)
+        # update the data source with markers
+        source.add(fmarkers, name='fmarkers')
 
-    # create an instance of the glyph
-    glyph_ax1 = markers[fid]()
-    glyph_ax1.update(x='x', y='y1', **glyph_opts)
+        p1 = ax1.scatter(x='x', y='y1', marker='fmarkers', source=source,
+                         line_width=0, angle=0.7, **glyph_opts)
 
-    p1 = ax1.add_glyph(source, glyph=glyph_ax1,
-                       nonselection_glyph=nonsel_glyph)
-    p1_err = errorbar(fig=ax1, x=source.data['x'], y=source.data['y1'],
-                      color=color, yerr=y1_err)
+        p2 = ax2.scatter(x='x', y='y2', marker='fmarkers', source=source,
+                         line_width=0, angle=0.7, **glyph_opts)
+    else:
+        p1 = ax1.scatter(x='x', y='y1', marker=markers[fid], source=source,
+                         line_width=0, **glyph_opts)
 
-    glyph_ax2 = markers[fid]()
-    glyph_ax2.update(x='x', y='y2', **glyph_opts)
-    p2 = ax2.add_glyph(source, glyph=glyph_ax2,
-                       nonselection_glyph=nonsel_glyph)
-    p2_err = errorbar(fig=ax2, x=source.data['x'], y=source.data['y2'],
-                      color=color, yerr=y2_err)
+        p2 = ax2.scatter(x='x', y='y2', marker=markers[fid], source=source,
+                         line_width=0, **glyph_opts)
+
+    # add a check for whether all the in y data were NaNs
+    # this causes the errorbars to fail if they all are
+    # they must be checked
+    if np.all(np.isnan(source.data['y1'])):
+        p1_err = errorbar(fig=ax1, x=source.data['x'], y=source.data['y1'],
+                          color=color, yerr=None)
+        p2_err = errorbar(fig=ax2, x=source.data['y2'], y=source.data['y2'],
+                          color=color, yerr=None)
+    else:
+        p1_err = errorbar(fig=ax1, x=source.data['x'], y=source.data['y1'],
+                          color=color, yerr=y1err)
+        p2_err = errorbar(fig=ax2, x=source.data['x'], y=source.data['y2'],
+                          color=color, yerr=y2err)
+
+    # link the visible properties of the two plots
+    p1.js_link('visible', p2, 'visible')
+    p2.js_link('visible', p1, 'visible')
+
+    if p1_err:
+        p1.glyph.js_link('size', p1_err, 'line_width')
+        p2.glyph.js_link('size', p2_err, 'line_width')
 
     return p1, p1_err, p2, p2_err
 
@@ -648,6 +658,7 @@ def ant_select_callback():
     code = """
             var i;
              //if toggle button active
+             //num_groups: number of chunks available
             if (this.active==false)
                 {
                     this.label='Select all Antennas';
@@ -668,7 +679,7 @@ def ant_select_callback():
 
                     }
 
-                    batchsel.active = [0,1,2,3]
+                    batchsel.active = [...Array(num_groups).keys()]
                 }
             """
 
@@ -680,18 +691,17 @@ def toggle_err_callback():
         Returns : string
     """
     code = """
-            var i;
+            let i;
              //if toggle button active
             if (this.active==false)
                 {
                     this.label='Show All Error bars';
 
-
                     for(i=0; i<err1.length; i++){
-                        err1[i][1][0].visible = false;
-                        //checking for error on phase and imaginary planes as these tend to go off
-                        if (err2[i][1][0]){
-                            err2[i][1][0].visible = false;
+                        //check if the error bars are present first
+                        if (err1[i]){
+                            err1[i].visible = false;
+                            err2[i].visible = false;
                         }
 
 
@@ -701,9 +711,12 @@ def toggle_err_callback():
             else{
                     this.label='Hide All Error bars';
                     for(i=0; i<err1.length; i++){
-                        err1[i][1][0].visible = true;
-                        if (err2[i][1][0]){
-                            err2[i][1][0].visible = true;
+                        //only switch on if corresponding plot is on
+                        if(ax1s[i].visible){
+                            if (err1[i]){
+                                err1[i].visible = true;
+                                err2[i].visible = true;
+                            }
                         }
                     }
                 }
@@ -742,8 +755,10 @@ def batch_select_callback():
                         //show all items in the active batch
 
                         for (f=0; f<nfields; f++){
-                            bax1[i][k][1][f].visible = true;
-                            bax2[i][k][1][f].visible = true;
+                            if (bax1[i][k]){
+                                bax1[i][k][1][f].visible = true;
+                                bax2[i][k][1][f].visible = true;
+                                }
                         }
                         k++;
 
@@ -754,8 +769,10 @@ def batch_select_callback():
                     k=0;
                     while (k < batch_size){
                         for (f=0; f<nfields; f++){
-                            bax1[i][k][1][f].visible = false;
-                            bax2[i][k][1][f].visible = false;
+                            if (bax1[i][k]){
+                                bax1[i][k][1][f].visible = false;
+                                bax2[i][k][1][f].visible = false;
+                                }
                         }
                         k++;
                     }
@@ -797,30 +814,10 @@ def legend_toggle_callback():
                     }
                 }
 
-
-
-                if (this.value == "elo"){
-                    for(i=0; i<len; i++){
-                        loax1_err[i].visible = true;
-                        loax2_err[i].visible = true;
-
-                    }
-                }
-
-                else{
-                    for(i=0; i<len; i++){
-                        loax1_err[i].visible = false;
-                        loax2_err[i].visible = false;
-
-                    }
-                }
-
                 if (this.value == "non"){
                     for(i=0; i<len; i++){
                         loax1[i].visible = false;
                         loax2[i].visible = false;
-                        loax1_err[i].visible = false;
-                        loax2_err[i].visible = false;
 
                     }
                 }
@@ -925,8 +922,10 @@ def flag_callback():
     code = """
             //sources: list of all different sources for the different antennas and available sources
             //n_sources: number of sources
+            //flagging: status of flag_data
 
             let n_sources =  sources.length;
+            let state = cb_obj.active.length;
 
             let init_src = Array();
             let src_1 = Array();
@@ -934,8 +933,8 @@ def flag_callback():
             for(item in sources){init_src[item] = sources[item][0];}
             for(item in sources){src_1[item] = sources[item][1];}
 
-            if (cb_obj.active){
-                cb_obj.label = 'Un-Flag';
+            if (state==1){
+                cb_obj.label = flagging ? 'Flag' : 'Un-Flag';
                 for (i=0; i<n_sources; i++){
                     init_src[i].data.y1 = src_1[i].data.iy1;
                     init_src[i].data.y2 = src_1[i].data.iy2;
@@ -943,7 +942,7 @@ def flag_callback():
                 }
             }
             else{
-                cb_obj.label = 'Flag';
+                cb_obj.label = flagging ? 'Un-Flag' : 'Flag';
                 for (i=0; i<n_sources; i++){
                     init_src[i].data.y1 = src_1[i].data.y1;
                     init_src[i].data.y2 = src_1[i].data.y2;
@@ -956,7 +955,7 @@ def flag_callback():
     return code
 
 
-def create_legend_batches(num_leg_objs, li_ax1, li_ax2, lierr_ax1, lierr_ax2, batch_size=16):
+def create_legend_batches(num_leg_objs, li_ax1, li_ax2, batch_size=16):
     """Automates creation of antenna batches of 16 each unless otherwise
 
         batch_0 : li_ax1[:16]
@@ -976,30 +975,21 @@ def create_legend_batches(num_leg_objs, li_ax1, li_ax2, lierr_ax1, lierr_ax2, ba
     li_ax2: list
                 List containing all legend items for antennas for 2nd figure
                 Items are in the form (antenna_legend, [glyph])
-    lierr_ax1: list
-                List containing legend items for errorbars for 1st figure
-                Items are in the form (error_legend, [glyph])
-    lierr_ax2: list
-                List containing legend items for errorbars for 2nd figure
-                Items are in the form (error_legend, [glyph])
-
     Outputs
     -------
 
-    (bax1, bax1_err, bax2, bax2_err): Tuple
+    (bax1, bax2: Tuple
                 Tuple containing List of lists which have batch_size number of legend items for each batch.
-                Results in batches for figure1 antenna legends, figure1 error legends, figure2 antenna legends, figure2 error legends.
+                Results in batches for figure1 antenna legends, figure2 antenna legends
 
                 e.g bax1 = [[batch0], [batch1], ...,  [batch_numOfBatches]]
 
     """
 
-    bax1, bax1_err, bax2, bax2_err = [], [], [], []
+    bax1, bax2 = [], []
 
     # condense the returned list if two fields were plotted
-    li_ax1, li_ax2, lierr_ax1, lierr_ax2 = list(map(condense_legend_items,
-                                                    [li_ax1, li_ax2,
-                                                     lierr_ax1, lierr_ax2]))
+    li_ax1, li_ax2 = list(map(condense_legend_items, [li_ax1, li_ax2]))
 
     j = 0
     for i in range(num_leg_objs):
@@ -1009,20 +999,17 @@ def create_legend_batches(num_leg_objs, li_ax1, li_ax2, lierr_ax1, lierr_ax2, ba
         if i == num_leg_objs:
             bax1.extend([li_ax1[j:]])
             bax2.extend([li_ax2[j:]])
-            bax1_err.extend([lierr_ax1[j:]])
-            bax2_err.extend([lierr_ax2[j:]])
+
         else:
             bax1.extend([li_ax1[j:j + batch_size]])
             bax2.extend([li_ax2[j:j + batch_size]])
-            bax1_err.extend([lierr_ax1[j:j + batch_size]])
-            bax2_err.extend([lierr_ax2[j:j + batch_size]])
 
         j += batch_size
 
-    return bax1, bax1_err, bax2, bax2_err
+    return bax1, bax2
 
 
-def create_legend_objs(num_leg_objs, bax1, baerr_ax1, bax2, baerr_ax2):
+def create_legend_objs(num_leg_objs, bax1, bax2):
     """Creates legend objects using items from batches list
        Legend objects allow legends be positioning outside the main plot
 
@@ -1034,15 +1021,10 @@ def create_legend_objs(num_leg_objs, bax1, baerr_ax1, bax2, baerr_ax2):
          Batches for antenna legends of 1st figure
    bax2: list
          Batches for antenna legends of 2nd figure
-   baerr_ax1: list
-         Batches for error bar legends of 1st figure
-   baerr_ax2: list
-         Batches for error bar legends of 2nd figure
-
 
    Outputs
    -------
-   (lo_ax1, loerr_ax1, lo_ax2, loerr_ax2) tuple
+   (lo_ax1, lo_ax2) tuple
             Tuple containing dictionaries with legend objects for
             ax1 antenna legend objects, ax1 error bar legend objects,
             ax2 antenna legend objects, ax2 error bar legend objects
@@ -1054,7 +1036,7 @@ def create_legend_objs(num_leg_objs, bax1, baerr_ax1, bax2, baerr_ax2):
                 items=batch_0, location='top_right', click_policy='hide')
     """
 
-    lo_ax1, lo_ax2, loerr_ax1, loerr_ax2 = {}, {}, {}, {}
+    lo_ax1, lo_ax2 = {}, {}
 
     l_opts = dict(click_policy='hide',
                   orientation='horizontal',
@@ -1063,16 +1045,13 @@ def create_legend_objs(num_leg_objs, bax1, baerr_ax1, bax2, baerr_ax2):
                   glyph_width=10)
 
     for i in range(num_leg_objs):
-        lo_ax1['leg_%s' % str(i)] = Legend(items=bax1[i],
-                                           **l_opts)
-        lo_ax2['leg_%s' % str(i)] = Legend(items=bax2[i],
-                                           **l_opts)
-        loerr_ax1['leg_%s' % str(i)] = Legend(items=baerr_ax1[i],
-                                              **l_opts)
-        loerr_ax2['leg_%s' % str(i)] = Legend(items=baerr_ax2[i],
-                                              **l_opts)
+        leg1 = Legend(items=bax1[i], **l_opts)
+        leg2 = Legend(items=bax2[i], **l_opts)
 
-    return lo_ax1, loerr_ax1, lo_ax2, loerr_ax2
+        lo_ax1['leg_%s' % str(i)] = leg1
+        lo_ax2['leg_%s' % str(i)] = leg2
+
+    return lo_ax1, lo_ax2
 
 
 def gen_checkbox_labels(batch_size, num_leg_objs, antnames):
@@ -1119,8 +1098,9 @@ def save_html(hname, plot_layout):
     -------
     Nothing
     """
-    output_file(hname + ".html")
-    output = save(plot_layout, hname + ".html", title=hname)
+    hname = hname + ".html"
+    output_file(hname)
+    output = save(plot_layout, hname, title=hname)
     # uncomment next line to automatically plot on web browser
     # show(layout)
 
@@ -1376,6 +1356,40 @@ def condense_legend_items(inlist):
     return outlist
 
 
+def gen_flag_data_markers(y, fid=None, markers=None, fmarker='circle_x'):
+    """Generate different markers for where data has been flagged
+
+        Inputs
+        ------
+        y: numpy.ndarray
+           The flagged data
+        fid: int
+             field id number to identify the marker to be used
+        markers: list
+                 A list of all available markers
+        fmarker: str
+                 the marker to be used for flagged data
+
+        Outputs
+        -------
+        masked_list: list
+                     Edited list containing markers for flagged data
+    """
+
+    # fill an array with the unflagged marker value
+    markers_arr = np.full(y.shape, fill_value=markers[fid], dtype='<U17')
+
+    # mask only where there are nan values
+    masked_markers_arr = np.ma.masked_where(np.isnan(y), markers_arr)
+    # fill with the different marker
+    masked_markers_arr.fill_value = fmarker
+
+    # return filled matrix
+    masked_list = masked_markers_arr.filled()
+
+    return masked_list
+
+
 def main(**kwargs):
     """Main function"""
     if len(kwargs) == 0:
@@ -1419,7 +1433,7 @@ def main(**kwargs):
         gain_types = kwargs.get('gain_types', [])
 
     # To flag or not
-    flag_data = False
+    flag_data = True
 
     # default spwid
     spwid = 0
@@ -1465,10 +1479,21 @@ def main(**kwargs):
         ncorrs = tt.FLAG.corr.data
 
         # convert field ids to strings
+        valid_fids = []
         if field_ids is None:
-            fids = [str(f) for f in fields.tolist()]
+            valid_fids = [str(f) for f in fields.tolist()]
         else:
-            fids = field_ids
+            for f in field_ids:
+                if f.isdigit():
+                    new_f = int(f)
+                else:
+                    new_f = name_2id(mytab, f)
+
+                if new_f in fields:
+                    valid_fids.append(new_f)
+                else:
+                    logger.info('Field {} not found in {}.'.format(f, mytab))
+                    continue
 
         freqs = (vu.get_frequencies(mytab, spwid=spwid) / GHZ).data.compute()
 
@@ -1510,28 +1535,19 @@ def main(**kwargs):
         # forming Legend object items for data and errors
         legend_items_ax1 = []
         legend_items_ax2 = []
-        legend_items_err_ax1 = []
-        legend_items_err_ax2 = []
+        ebars_ax1 = []
+        ebars_ax2 = []
 
         stats_ax1 = []
         stats_ax2 = []
 
         sources = []
+
         # enumerating available field ids incase of large fids
-        for enum_fid, field in enumerate(fids):
+        for enum_fid, field in enumerate(valid_fids):
 
-            if field.isdigit():
-                field = int(field)
-            else:
-                field = name_2id(mytab, field)
-
-            if int(field) not in fields.tolist():
-                logger.info(
-                    'Skipping table: {} : Field id {} not found.'.format(mytab, field))
-                continue
-
-            stats_text = stats_display(
-                mytab, gain_type, doplot, corr, field, flag=flag_data)
+            stats_text = stats_display(mytab, gain_type, doplot, corr,
+                                       field, flag=flag_data)
             stats_ax1.append(stats_text[0])
             stats_ax2.append(stats_text[1])
 
@@ -1550,7 +1566,8 @@ def main(**kwargs):
 
                 subtab = newtab.where(newtab.ANTENNA1 == int(ant), drop=True)
 
-                # To Do: Have a flag option in cmdline
+                # depending on the status of flag_data, this may
+                # be either flagged or unflagged data
                 data_obj = DataCoreProcessor(subtab, mytab, gain_type,
                                              fid=field, antenna=ant,
                                              doplot=doplot, corr=corr,
@@ -1567,7 +1584,7 @@ def main(**kwargs):
                 y2_err = ready_data.y2_err
                 y2label = ready_data.y2_label
 
-                # inverse flag data object
+                # inverse data object
                 infl_data_obj = DataCoreProcessor(subtab, mytab, gain_type,
                                                   fid=field, antenna=ant,
                                                   doplot=doplot, corr=corr,
@@ -1628,7 +1645,7 @@ def main(**kwargs):
 
                 p1, p1_err, p2, p2_err = make_plots(
                     source=source, color=y1col, ax1=ax1, ax2=ax2,
-                    fid=enum_fid, y1_err=y1_err, y2_err=y2_err)
+                    fid=enum_fid, y1err=y1_err, y2err=y2_err)
 
                 # hide all the other plots until legend is clicked
                 if ant > 0:
@@ -1642,8 +1659,8 @@ def main(**kwargs):
                 legend_items_ax1.append((legend, [p1]))
                 legend_items_ax2.append((legend, [p2]))
                 # for the errors
-                legend_items_err_ax1.append((legend_err, [p1_err]))
-                legend_items_err_ax2.append((legend_err, [p2_err]))
+                ebars_ax1.append(p1_err)
+                ebars_ax2.append(p2_err)
 
                 subtab.close()
 
@@ -1666,26 +1683,19 @@ def main(**kwargs):
         # for each plot
         num_legend_objs = int(np.ceil(len(plotants) / BATCH_SIZE))
 
-        batches_ax1, batches_ax1_err, batches_ax2, batches_ax2_err = \
-            create_legend_batches(num_legend_objs, legend_items_ax1,
-                                  legend_items_ax2, legend_items_err_ax1,
-                                  legend_items_err_ax2,
-                                  batch_size=BATCH_SIZE)
+        batches_ax1, batches_ax2 = create_legend_batches(num_legend_objs,
+                                                         legend_items_ax1,
+                                                         legend_items_ax2,
+                                                         batch_size=BATCH_SIZE)
 
-        legend_objs_ax1, legend_objs_ax1_err, legend_objs_ax2,\
-            legend_objs_ax2_err = create_legend_objs(num_legend_objs,
-                                                     batches_ax1,
-                                                     batches_ax1_err,
-                                                     batches_ax2,
-                                                     batches_ax2_err)
+        legend_objs_ax1, legend_objs_ax2 = create_legend_objs(num_legend_objs,
+                                                              batches_ax1,
+                                                              batches_ax2)
 
         # adding legend objects to the layouts
         for i in range(num_legend_objs):
             ax1.add_layout(legend_objs_ax1['leg_%s' % str(i)], 'below')
             ax2.add_layout(legend_objs_ax2['leg_%s' % str(i)], 'below')
-
-            ax1.add_layout(legend_objs_ax1_err['leg_%s' % str(i)], 'below')
-            ax2.add_layout(legend_objs_ax2_err['leg_%s' % str(i)], 'below')
 
         # adding plot titles
         ax2.add_layout(ax2_title, 'above')
@@ -1714,11 +1724,11 @@ def main(**kwargs):
         # Dropdown to hide and show legends
         legend_toggle = Select(title="Showing Legends: ", value="non",
                                options=[("alo", "Antennas"),
-                                        ("elo", "Errors"), ("non", "None")],
+                                        ("non", "None")],
                                width=150, height=45)
 
-        # creating size slider for the plots
-        size_slider = Slider(end=15, start=1, step=0.5,
+        # creating glyph size slider for the plots
+        size_slider = Slider(end=15, start=0.4, step=0.1,
                              value=4, title='Glyph size',
                              **w_dims)
 
@@ -1732,10 +1742,11 @@ def main(**kwargs):
 
         try:
             field_labels = ["Field {} {}".format(fnames[int(x)],
-                                                 fsyms[enum_fid]) for x in fields]
+                                                 fsyms[enum_fid]) for enum_fid, x in enumerate(fields)]
         except UnicodeEncodeError:
             field_labels = ["Field {} {}".format(fnames[int(x)],
-                                                 fsyms[enum_fid].encode('utf-8')) for x in fields]
+                                                 fsyms[enum_fid].encode('utf-8')) for enum_fid, x in enumerate(fields)]
+
         field_selector = CheckboxGroup(labels=field_labels,
                                        active=fields.tolist(),
                                        **w_dims)
@@ -1745,8 +1756,10 @@ def main(**kwargs):
         title_fontslider = Slider(end=35, start=10, step=1, value=15,
                                   title='Title size', **w_dims)
 
-        toggle_flag = Toggle(label='Flag', button_type='warning',
-                             active=flag_data, **w_dims)
+        # if flag_data is true, i.e data is flagged, label==Un-flag,
+        # button==inactive [not flag_data], otherwise button==in
+        toggle_flag = CheckboxGroup(labels=['Show Flagged-out Data'],
+                                    active=[], **w_dims)
 
         ######################################################################
         ############## Defining widget Callbacks ############################
@@ -1754,63 +1767,65 @@ def main(**kwargs):
 
         ant_select.callback = CustomJS(args=dict(glyph1=legend_items_ax1,
                                                  glyph2=legend_items_ax2,
-                                                 batchsel=batch_select),
+                                                 batchsel=batch_select,
+                                                 num_groups=num_legend_objs),
                                        code=ant_select_callback())
 
-        toggle_err.callback = CustomJS(args=dict(err1=legend_items_err_ax1,
-                                                 err2=legend_items_err_ax2),
-                                       code=toggle_err_callback())
+        toggle_err.js_on_click(CustomJS(args=dict(ax1s=ax1_plots,
+                                                  err1=ebars_ax1,
+                                                  err2=ebars_ax2),
+                                        code=toggle_err_callback()))
 
         # BATCH SELECTION
         batch_select.callback = CustomJS(
             args=dict(bax1=batches_ax1,
-                      bax1_err=batches_ax1_err,
                       bax2=batches_ax2,
-                      bax2_err=batches_ax2_err,
                       batch_size=BATCH_SIZE,
                       antsel=ant_select),
             code=batch_select_callback())
 
         legend_toggle.callback = CustomJS(
-            args=dict(
-                loax1=listvalues(legend_objs_ax1),
-                loax1_err=listvalues(legend_objs_ax1_err),
-                loax2=listvalues(legend_objs_ax2),
-                loax2_err=listvalues(legend_objs_ax2_err)),
+            args=dict(loax1=listvalues(legend_objs_ax1),
+                      loax2=listvalues(legend_objs_ax2)),
             code=legend_toggle_callback())
 
-        size_slider.callback = CustomJS(args={'slide': size_slider,
-                                              'p1': ax1_plots,
-                                              'p2': ax2_plots},
-                                        code=size_slider_callback())
-        alpha_slider.callback = CustomJS(args={'alpha': alpha_slider,
-                                               'p1': ax1_plots,
-                                               'p2': ax2_plots},
-                                         code=alpha_slider_callback())
+        size_slider.js_on_change('value',
+                                 CustomJS(args={'slide': size_slider,
+                                                'p1': ax1_plots,
+                                                'p2': ax2_plots},
+                                          code=size_slider_callback()))
+        alpha_slider.js_on_change('value',
+                                  CustomJS(args={'alpha': alpha_slider,
+                                                 'p1': ax1_plots,
+                                                 'p2': ax2_plots},
+                                           code=alpha_slider_callback()))
         field_selector.callback = CustomJS(args={'fselect': field_selector,
                                                  'p1': ax1_plots,
                                                  'p2': ax2_plots,
                                                  'ants': plotants},
                                            code=field_selector_callback())
-        axis_fontslider.callback = CustomJS(args=dict(ax1=ax1.axis,
-                                                      ax2=ax2.axis),
-                                            code=axis_fs_callback())
-        title_fontslider.callback = CustomJS(args=dict(ax1=ax1.above,
-                                                       ax2=ax2.above),
-                                             code=title_fs_callback())
+        axis_fontslider.js_on_change('value',
+                                     CustomJS(args=dict(ax1=ax1.axis,
+                                                        ax2=ax2.axis),
+                                              code=axis_fs_callback()))
+        title_fontslider.js_on_change('value',
+                                      CustomJS(args=dict(ax1=ax1.above,
+                                                         ax2=ax2.above),
+                                               code=title_fs_callback()))
 
         toggle_flag.callback = CustomJS(args=dict(sources=sources,
-                                                  nants=plotants),
+                                                  nants=plotants,
+                                                  flagging=flag_data),
                                         code=flag_callback())
 
         #################################################################
         ########## Define widget layouts #################################
         ##################################################################
 
-        a = row([ant_select, toggle_err, toggle_flag, size_slider,
+        a = row([ant_select, toggle_err, legend_toggle, size_slider,
                  alpha_slider])
-        b = row([batch_select, field_selector, axis_fontslider,
-                 title_fontslider, legend_toggle, ])
+        b = row([toggle_flag, batch_select, field_selector, axis_fontslider,
+                 title_fontslider])
 
         plot_widgets = widgetbox([a, b], sizing_mode='scale_both')
 
@@ -1827,7 +1842,9 @@ def main(**kwargs):
                               **grid_specs)
 
         final_layout.append(layout)
-        logger.info("Table {} done.".format(mytab))
+
+    logger.info("Table {} done.".format(mytab))
+
     if image_name:
         save_svg_image(image_name, ax1, ax2,
                        legend_items_ax1, legend_items_ax2)
@@ -1846,7 +1863,7 @@ def main(**kwargs):
                 mytab = datetime.now().strftime('%Y%m%d_%H%M%S')
             html_name = "{}_corr_{}_{}_field_{}".format(mytab, corr,
                                                         doplot,
-                                                        ''.join(fids))
+                                                        ''.join(valid_fids))
             save_html(html_name, final_layout)
 
         logger.info("Rendered: {}.html".format(html_name))